Correction of typos, minor changes and translation
39: 
https://prnt.sc/12k9mt8
In 3 lines, it looks terrible. Nothing has changed in the context, but it will look better
98:
https://prnt.sc/12k9spv
The hyphenation of the name looks terrible
118:
https://prnt.sc/12k9zhy
Same thing as with 98
260:
https://prnt.sc/12ka2xa
The transfer of one letter looks just disgusting. It is better to simply delete it. There will be a feeling that the girl's head is blocking the word.
314, 688, 728:
Just little remark
1352:
It was necessary to translate "Damage multiplier". Translated, literally "Health multiplier". Now the translation is correct
1436:
It was necessary to translate "Maximum health capacity". Translated, literally "The player's health count". Now the translation is correct
822, 826, 832, 836, 842, 846, 1211, 1215, 1447, 1455, 1462, 1467, 1472, 1478, 1490, 1497, 1501:
Translation into Russian