Update Tips.txt
Added the following tips:

If you need BTC, try selling unwanted items to the Weaponsmith.  The "Bargaining" stat helps with this.
Inventory getting full?  Lost BTC when you flatlined?  Store items or BTC in the [BANK OF ARASAKA].

Calibrating gear makes it stronger, but calibrated gear cannot be gifted or sold.
Each level of calibration on gear adds more to the item than the last level, but is more likely to break the item as well.

Looking for the tutorial?  It's in Profile -> Tutorial.