Update de.json
ummm yes

Hi
 
My name is DragonSinns. I'm from Germany,
I downloaded your app a few days ago and was really impressed

what you did there all by yourself is really amazing

respect to you

Your German translation is not the yellow of the egg. But because of your fighting spirit and your outstanding achievements that you have to manage all by yourself up to now. I sat for 4 days and read over 10,000 lines so that you also have a chance on the German market
(On the Google Play Store, everyone complains about the "terrible translation")

I'm not (we say in Germany) a German grammar expert and there are probably still 100-150 mistakes that need to be corrected
But I hope you appreciate my work

The whole thing has only one catch

I couldn't translate some of the lines, or only partially, because I either didn't know what the text's function was or there was no German word for it
e.g. the entire ItemTrait

but with a little help from you
I can translate all these remaining lines
[ I do not give up :-) ]

if you need help I'm happy to be there for you

(I hope it's okay that I just use Google to translate this message. It's 11:00 PM in Germany and I want to go to bed)

Greetings DragonSinns