Update global_skip.md
Nobody waits 1 minute anymore due to the sheer player count. This used to be a thing when the game had... 300 peak players max (100 average) but now the queue time is so long most (probably over 95%) people can immediately go back to their afk tasks and not lose out on any skips.

Added some info just so we don't just have a 1 liner. It feels kinda eh that way.