Update id.json
As a native speaker of the Indonesian language who deeply passionate about its preservation and development, I am committed to contribute in the successful execution of this project and contributing to the long-term sustainability of the Indonesian language ecosystem. I make some minor change in order to maximize gameplay experience. I will propose the next update as soon as possible and i hope you accept the changes i made.  :)